% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{add_permission}
\alias{add_permission}
\title{Adds a permission to the resource policy associated with the specified AWS Lambda function}
\usage{
add_permission(FunctionName, StatementId, Action, Principal, SourceArn = NULL,
  SourceAccount = NULL, EventSourceToken = NULL, Qualifier = NULL,
  RevisionId = NULL)
}
\arguments{
\item{FunctionName}{The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{StatementId}{A unique statement identifier.}

\item{Action}{The AWS Lambda action you want to allow in this statement. Each Lambda action is a string starting with \code{lambda:} followed by the API name . For example, \code{lambda:CreateFunction}. You can use wildcard (\code{lambda:*}) to grant permission for all AWS Lambda actions.}

\item{Principal}{The principal who is getting this permission. The principal can be an AWS service (e.g. \code{s3.amazonaws.com} or \code{sns.amazonaws.com}) for service triggers, or an account ID for cross-account access. If you specify a service as a principal, use the \code{SourceArn} parameter to limit who can invoke the function through that service.}

\item{SourceArn}{The Amazon Resource Name of the invoker.

If you add a permission to a service principal without providing the source ARN, any AWS account that creates a mapping to your function ARN can invoke your Lambda function.}

\item{SourceAccount}{This parameter is used for S3 and SES. The AWS account ID (without a hyphen) of the source owner. For example, if the \code{SourceArn} identifies a bucket, then this is the bucket owner's account ID. You can use this additional condition to ensure the bucket you specify is owned by a specific account (it is possible the bucket owner deleted the bucket and some other AWS account created the bucket). You can also use this condition to specify all sources (that is, you don't specify the \code{SourceArn}) owned by a specific account.}

\item{EventSourceToken}{A unique token that must be supplied by the principal invoking the function. This is currently only used for Alexa Smart Home functions.}

\item{Qualifier}{Specify a version or alias to add permissions to a published version of the function.}

\item{RevisionId}{An optional value you can use to ensure you are updating the latest update of the function version or alias. If the \code{RevisionID} you pass doesn't match the latest \code{RevisionId} of the function or alias, it will fail with an error message, advising you to retrieve the latest function version or alias \code{RevisionID} using either GetFunction or GetAlias}
}
\description{
Adds a permission to the resource policy associated with the specified AWS Lambda function. You use resource policies to grant permissions to event sources that use the \emph{push} model. In a \emph{push} model, event sources (such as Amazon S3 and custom applications) invoke your Lambda function. Each permission you add to the resource policy allows an event source permission to invoke the Lambda function.
}
\details{
Permissions apply to the Amazon Resource Name (ARN) used to invoke the function, which can be unqualified (the unpublished version of the function), or include a version or alias. If a client uses a version or alias to invoke a function, use the \code{Qualifier} parameter to apply permissions to that ARN. For more information about versioning, see \href{http://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html}{AWS Lambda Function Versioning and Aliases}.

This operation requires permission for the \code{lambda:AddPermission} action.
}
\examples{
# This example adds a permission for an S3 bucket to invoke a Lambda
# function.
add_permission(
  Action = "lambda:InvokeFunction",
  FunctionName = "MyFunction",
  Principal = "s3.amazonaws.com",
  SourceAccount = "123456789012",
  SourceArn = "arn:aws:s3:::examplebucket/*",
  StatementId = "ID-1"
)

}
