% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{create_alias}
\alias{create_alias}
\title{Creates an alias for a Lambda function version}
\usage{
create_alias(FunctionName, Name, FunctionVersion, Description = NULL,
  RoutingConfig = NULL)
}
\arguments{
\item{FunctionName}{[required] The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Name}{[required] The name of the alias.}

\item{FunctionVersion}{[required] The function version that the alias invokes.}

\item{Description}{A description of the alias.}

\item{RoutingConfig}{The \href{http://docs.aws.amazon.com/lambda/latest/dg/lambda-traffic-shifting-using-aliases.html}{routing configuration} of the alias.}
}
\description{
Creates an \href{http://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html}{alias} for a Lambda function version. Use aliases to provide clients with a function identifier that you can update to invoke a different version.
}
\details{
You can also map an alias to split invocation requests between two versions. Use the \code{RoutingConfig} parameter to specify a second version and the percentage of invocation requests that it receives.
}
\section{Accepted Parameters}{
\preformatted{create_alias(
  FunctionName = "string",
  Name = "string",
  FunctionVersion = "string",
  Description = "string",
  RoutingConfig = list(
    AdditionalVersionWeights = list(
      123.0
    )
  )
)
}
}

