% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{delete_alias}
\alias{delete_alias}
\title{Deletes a Lambda function alias}
\usage{
delete_alias(FunctionName, Name)
}
\arguments{
\item{FunctionName}{[required] The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Name}{[required] The name of the alias.}
}
\description{
Deletes a Lambda function \href{http://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html}{alias}.
}
\section{Accepted Parameters}{
\preformatted{delete_alias(
  FunctionName = "string",
  Name = "string"
)
}
}

\examples{
# This operation deletes a Lambda function alias
\donttest{delete_alias(
  FunctionName = "myFunction",
  Name = "alias"
)}

}
