% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{delete_event_source_mapping}
\alias{delete_event_source_mapping}
\title{Deletes an event source mapping}
\usage{
delete_event_source_mapping(
  UUID = "string"
)
}
\arguments{
\item{UUID}{[required] The identifier of the event source mapping.}
}
\description{
Deletes an event source mapping.
}
\examples{
# This operation deletes a Lambda function event source mapping
delete_event_source_mapping(
  UUID = "12345kxodurf3443"
)

}
