% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{delete_layer_version}
\alias{delete_layer_version}
\title{Deletes a version of a function layer}
\usage{
delete_layer_version(
  LayerName = "string",
  VersionNumber = 123
)
}
\arguments{
\item{LayerName}{[required] The name of the layer.}

\item{VersionNumber}{[required] The version number.}
}
\description{
Deletes a version of a function layer. Deleted versions can no longer be viewed or added to functions. However, a copy of the version remains in Lambda until no functions refer to it.
}
\examples{

}
