% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{get_alias}
\alias{get_alias}
\title{Returns the specified alias information such as the alias ARN, description, and function version it is pointing to}
\usage{
get_alias(FunctionName, Name)
}
\arguments{
\item{FunctionName}{The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Name}{Name of the alias for which you want to retrieve information.}
}
\description{
Returns the specified alias information such as the alias ARN, description, and function version it is pointing to. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/aliases-intro.html}{Introduction to AWS Lambda Aliases}.
}
\details{
This requires permission for the \code{lambda:GetAlias} action.
}
\examples{
# This operation retrieves a Lambda function alias
get_alias(
  FunctionName = "myFunction",
  Name = "myFunctionAlias"
)

}
