% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{get_function}
\alias{get_function}
\title{Returns the configuration information of the Lambda function and a presigned URL link to the}
\usage{
get_function(FunctionName, Qualifier = NULL)
}
\arguments{
\item{FunctionName}{[required] The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Qualifier}{Specify a version or alias to get details about a published version of the function.}
}
\description{
Returns the configuration information of the Lambda function and a presigned URL link to the .zip file you uploaded with CreateFunction so you can download the .zip file. Note that the URL is valid for up to 10 minutes. The configuration information is the same information you provided as parameters when uploading the function.
}
\details{
Use the \code{Qualifier} parameter to retrieve a published version of the function. Otherwise, returns the unpublished version (\code{$LATEST}). For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html}{AWS Lambda Function Versioning and Aliases}.

This operation requires permission for the \code{lambda:GetFunction} action.
}
\section{Accepted Parameters}{
\preformatted{get_function(
  FunctionName = "string",
  Qualifier = "string"
)
}
}

\examples{
# This operation retrieves a Lambda function's event source mapping
\donttest{get_function(
  FunctionName = "myFunction",
  Qualifier = "1"
)}

}
