% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{get_function_configuration}
\alias{get_function_configuration}
\title{Returns the configuration information of the Lambda function}
\usage{
get_function_configuration(FunctionName, Qualifier = NULL)
}
\arguments{
\item{FunctionName}{[required] The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Qualifier}{Specify a version or alias to get details about a published version of the function.}
}
\description{
Returns the configuration information of the Lambda function. This the same information you provided as parameters when uploading the function by using CreateFunction.
}
\details{
If you are using the versioning feature, you can retrieve this information for a specific function version by using the optional \code{Qualifier} parameter and specifying the function version or alias that points to it. If you don't provide it, the API returns information about the \$LATEST version of the function. For more information about versioning, see \href{http://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html}{AWS Lambda Function Versioning and Aliases}.

This operation requires permission for the \code{lambda:GetFunctionConfiguration} operation.
}
\section{Accepted Parameters}{
\preformatted{get_function_configuration(
  FunctionName = "string",
  Qualifier = "string"
)
}
}

\examples{
# This operation retrieves a Lambda function's event source mapping
\donttest{get_function_configuration(
  FunctionName = "myFunction",
  Qualifier = "1"
)}

}
