% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{get_layer_version}
\alias{get_layer_version}
\title{Returns information about a version of a function layer, with a link to download the layer archive that's valid for 10 minutes}
\usage{
get_layer_version(
  LayerName = "string",
  VersionNumber = 123
)
}
\arguments{
\item{LayerName}{[required] The name of the layer.}

\item{VersionNumber}{[required] The version number.}
}
\description{
Returns information about a version of a function layer, with a link to download the layer archive that's valid for 10 minutes.
}
\examples{

}
