% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{get_layer_version_policy}
\alias{get_layer_version_policy}
\title{Returns the permission policy for a layer version}
\usage{
get_layer_version_policy(
  LayerName = "string",
  VersionNumber = 123
)
}
\arguments{
\item{LayerName}{[required] The name of the layer.}

\item{VersionNumber}{[required] The version number.}
}
\description{
Returns the permission policy for a layer version. For more information, see AddLayerVersionPermission.
}
\examples{

}
