% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{get_policy}
\alias{get_policy}
\title{Returns the resource policy associated with the specified Lambda function}
\usage{
get_policy(FunctionName, Qualifier = NULL)
}
\arguments{
\item{FunctionName}{The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Qualifier}{You can specify this optional query parameter to specify a function version or an alias name in which case this API will return all permissions associated with the specific qualified ARN. If you don't provide this parameter, the API will return permissions that apply to the unqualified function ARN.}
}
\description{
Returns the resource policy associated with the specified Lambda function.
}
\details{
This action requires permission for the \code{lambda:GetPolicy action.}
}
\examples{
# This operation retrieves a Lambda function policy
get_policy(
  FunctionName = "myFunction",
  Qualifier = "1"
)

}
