% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{invoke}
\alias{invoke}
\title{Invokes a Lambda function}
\usage{
invoke(FunctionName, InvocationType = NULL, LogType = NULL,
  ClientContext = NULL, Payload = NULL, Qualifier = NULL)
}
\arguments{
\item{FunctionName}{The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{InvocationType}{Choose from the following options.
\itemize{
\item \code{RequestResponse} (default) - Invoke the function synchronously. Keep the connection open until the function returns a response or times out.
\item \code{Event} - Invoke the function asynchronously. Send events that fail multiple times to the function's dead-letter queue (if configured).
\item \code{DryRun} - Validate parameter values and verify that the user or role has permission to invoke the function.
}}

\item{LogType}{You can set this optional parameter to \code{Tail} in the request only if you specify the \code{InvocationType} parameter with value \code{RequestResponse}. In this case, AWS Lambda returns the base64-encoded last 4 KB of log data produced by your Lambda function in the \code{x-amz-log-result} header.}

\item{ClientContext}{Using the \code{ClientContext} you can pass client-specific information to the Lambda function you are invoking. You can then process the client information in your Lambda function as you choose through the context variable. For an example of a \code{ClientContext} JSON, see \href{http://docs.aws.amazon.com/mobileanalytics/latest/ug/PutEvents.html}{PutEvents} in the \emph{Amazon Mobile Analytics API Reference and User Guide}.

The ClientContext JSON must be base64-encoded and has a maximum size of 3583 bytes.

\code{ClientContext} information is returned only if you use the synchronous (\code{RequestResponse}) invocation type.}

\item{Payload}{JSON that you want to provide to your Lambda function as input.}

\item{Qualifier}{Specify a version or alias to invoke a published version of the function.}
}
\description{
Invokes a Lambda function. For an example, see \href{http://docs.aws.amazon.com/lambda/latest/dg/with-dynamodb-create-function.html#with-dbb-invoke-manually}{Create the Lambda Function and Test It Manually}.
}
\details{
Specify just a function name to invoke the latest version of the function. To invoke a published version, use the \code{Qualifier} parameter to specify a \href{http://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html}{version or alias}.

If you use the \code{RequestResponse} (synchronous) invocation option, the function will be invoked only once. If you use the \code{Event} (asynchronous) invocation option, the function will be invoked at least once in response to an event and the function must be idempotent to handle this.

For functions with a long timeout, your client may be disconnected during synchronous invocation while it waits for a response. Configure your HTTP client, SDK, firewall, proxy, or operating system to allow for long connections with timeout or keep-alive settings.

This operation requires permission for the \code{lambda:InvokeFunction} action.

The \code{TooManyRequestsException} noted below will return the following: \code{ConcurrentInvocationLimitExceeded} will be returned if you have no functions with reserved concurrency and have exceeded your account concurrent limit or if a function without reserved concurrency exceeds the account's unreserved concurrency limit. \code{ReservedFunctionConcurrentInvocationLimitExceeded} will be returned when a function with reserved concurrency exceeds its configured concurrency limit.
}
\examples{
# This operation invokes a Lambda function
invoke(
  ClientContext = "MyApp",
  FunctionName = "MyFunction",
  InvocationType = "Event",
  LogType = "Tail",
  Payload = "fileb://file-path/input.json",
  Qualifier = "1"
)

}
