% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{invoke_async}
\alias{invoke_async}
\title{For asynchronous function invocation, use Invoke}
\usage{
invoke_async(FunctionName, InvokeArgs)
}
\arguments{
\item{FunctionName}{The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{InvokeArgs}{JSON that you want to provide to your Lambda function as input.}
}
\description{
For asynchronous function invocation, use Invoke.
}
\details{
Submits an invocation request to AWS Lambda. Upon receiving the request, Lambda executes the specified function asynchronously. To see the logs generated by the Lambda function execution, see the CloudWatch Logs console.

This operation requires permission for the \code{lambda:InvokeFunction} action.
}
\examples{
# This operation invokes a Lambda function asynchronously
invoke_async(
  FunctionName = "myFunction",
  InvokeArgs = "fileb://file-path/input.json"
)

}
