% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{list_event_source_mappings}
\alias{list_event_source_mappings}
\title{Returns a list of event source mappings you created using the CreateEventSourceMapping (see CreateEventSourceMapping)}
\usage{
list_event_source_mappings(EventSourceArn = NULL, FunctionName = NULL,
  Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{EventSourceArn}{The Amazon Resource Name (ARN) of the Amazon Kinesis or DynamoDB stream. (This parameter is optional.)}

\item{FunctionName}{The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Version or Alias ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction:PROD}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Marker}{Optional string. An opaque pagination token returned from a previous \code{ListEventSourceMappings} operation. If present, specifies to continue the list from where the returning call left off.}

\item{MaxItems}{Optional integer. Specifies the maximum number of event sources to return in response. This value must be greater than 0.}
}
\description{
Returns a list of event source mappings you created using the \code{CreateEventSourceMapping} (see CreateEventSourceMapping).
}
\details{
For each mapping, the API returns configuration information. You can optionally specify filters to retrieve specific event source mappings.

This operation requires permission for the \code{lambda:ListEventSourceMappings} action.
}
\section{Accepted Parameters}{
\preformatted{list_event_source_mappings(
  EventSourceArn = "string",
  FunctionName = "string",
  Marker = "string",
  MaxItems = 123
)
}
}

