% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{list_versions_by_function}
\alias{list_versions_by_function}
\title{Lists all versions of a function}
\usage{
list_versions_by_function(
  FunctionName = "string",
  Marker = "string",
  MaxItems = 123
)
}
\arguments{
\item{FunctionName}{[required] The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Marker}{Optional string. An opaque pagination token returned from a previous \code{ListVersionsByFunction} operation. If present, indicates where to continue the listing.}

\item{MaxItems}{Optional integer. Specifies the maximum number of AWS Lambda function versions to return in response. This parameter value must be greater than 0.}
}
\description{
Lists all versions of a function. For information about versioning, see \href{http://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html}{AWS Lambda Function Versioning and Aliases}.
}
\examples{
# This operation retrieves a Lambda function versions
list_versions_by_function(
  FunctionName = "myFunction",
  Marker = "",
  MaxItems = 123L
)

}
