% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_package.R
\docType{package}
\name{paws.lambda-package}
\alias{paws.lambda}
\alias{paws.lambda-package}
\title{paws.lambda: AWS Lambda}
\description{
This is the *AWS Lambda API Reference*. The AWS Lambda
Developer Guide provides additional information. For the service
overview, see [What is AWS
Lambda](http://docs.aws.amazon.com/lambda/latest/dg/welcome.html), and
for information about how the service works, see [AWS Lambda: How it
Works](http://docs.aws.amazon.com/lambda/latest/dg/lambda-introduction.html)
in the **AWS Lambda Developer Guide**.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
