% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{publish_version}
\alias{publish_version}
\title{Creates a version from the current code and configuration of a function}
\usage{
publish_version(FunctionName, CodeSha256 = NULL, Description = NULL,
  RevisionId = NULL)
}
\arguments{
\item{FunctionName}{[required] The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{CodeSha256}{Only publish a version if the hash matches the value specified. Use this option to avoid publishing a version if the function code has changed since you last updated it. You can get the hash for the version you uploaded from the output of UpdateFunctionCode.}

\item{Description}{Specify a description for the version to override the description in the function configuration.}

\item{RevisionId}{Only update the function if the revision ID matches the ID specified. Use this option to avoid publishing a version if the function configuration has changed since you last updated it.}
}
\description{
Creates a \href{http://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html}{version} from the current code and configuration of a function. Use versions to create a snapshot of your function code and configuration that doesn't change.
}
\details{
AWS Lambda does not publish a version if the function's configuration and code hasn't changed since the last version. Use UpdateFunctionCode or UpdateFunctionConfiguration to update the function prior to publishing a version.

Clients can invoke versions directly or with an alias. To create an alias, use CreateAlias.
}
\section{Accepted Parameters}{
\preformatted{publish_version(
  FunctionName = "string",
  CodeSha256 = "string",
  Description = "string",
  RevisionId = "string"
)
}
}

\examples{
# This operation publishes a version of a Lambda function
\donttest{publish_version(
  CodeSha256 = "",
  Description = "",
  FunctionName = "myFunction"
)}

}
