% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes tags from a function}
\usage{
untag_resource(Resource, TagKeys)
}
\arguments{
\item{Resource}{[required] The function's Amazon Resource Name (ARN).}

\item{TagKeys}{[required] A list of tag keys to remove from the function.}
}
\description{
Removes \href{http://docs.aws.amazon.com/lambda/latest/dg/tagging.html}{tags} from a function.
}
\section{Accepted Parameters}{
\preformatted{untag_resource(
  Resource = "string",
  TagKeys = list(
    "string"
  )
)
}
}

