% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{update_alias}
\alias{update_alias}
\title{Using this API you can update the function version to which the alias points and the alias description}
\usage{
update_alias(FunctionName, Name, FunctionVersion = NULL,
  Description = NULL, RoutingConfig = NULL, RevisionId = NULL)
}
\arguments{
\item{FunctionName}{[required] The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Name}{[required] The alias name.}

\item{FunctionVersion}{Using this parameter you can change the Lambda function version to which the alias points.}

\item{Description}{You can change the description of the alias using this parameter.}

\item{RoutingConfig}{Specifies an additional version your alias can point to, allowing you to dictate what percentage of traffic will invoke each version. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/lambda-traffic-shifting-using-aliases.html}{Traffic Shifting Using Aliases}.}

\item{RevisionId}{An optional value you can use to ensure you are updating the latest update of the function version or alias. If the \code{RevisionID} you pass doesn't match the latest \code{RevisionId} of the function or alias, it will fail with an error message, advising you retrieve the latest function version or alias \code{RevisionID} using either GetFunction or GetAlias.}
}
\description{
Using this API you can update the function version to which the alias points and the alias description. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/aliases-intro.html}{Introduction to AWS Lambda Aliases}.
}
\details{
This requires permission for the lambda:UpdateAlias action.
}
\section{Accepted Parameters}{
\preformatted{update_alias(
  FunctionName = "string",
  Name = "string",
  FunctionVersion = "string",
  Description = "string",
  RoutingConfig = list(
    AdditionalVersionWeights = list(
      123.0
    )
  ),
  RevisionId = "string"
)
}
}

\examples{
# This operation updates a Lambda function alias
\donttest{update_alias(
  Description = "",
  FunctionName = "myFunction",
  FunctionVersion = "1",
  Name = "functionAlias"
)}

}
