% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{update_event_source_mapping}
\alias{update_event_source_mapping}
\title{You can update an event source mapping}
\usage{
update_event_source_mapping(UUID, FunctionName = NULL, Enabled = NULL,
  BatchSize = NULL)
}
\arguments{
\item{UUID}{[required] The event source mapping identifier.}

\item{FunctionName}{The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Version or Alias ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction:PROD}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Enabled}{Specifies whether AWS Lambda should actively poll the stream or not. If disabled, AWS Lambda will not poll the stream.}

\item{BatchSize}{The largest number of records that AWS Lambda will retrieve from your event source at the time of invoking your function. Your function receives an event with all the retrieved records.}
}
\description{
You can update an event source mapping. This is useful if you want to change the parameters of the existing mapping without losing your position in the stream. You can change which function will receive the stream records, but to change the stream itself, you must create a new mapping.
}
\details{
If you disable the event source mapping, AWS Lambda stops polling. If you enable again, it will resume polling from the time it had stopped polling, so you don't lose processing of any records. However, if you delete event source mapping and create it again, it will reset.

This operation requires permission for the \code{lambda:UpdateEventSourceMapping} action.
}
\section{Accepted Parameters}{
\preformatted{update_event_source_mapping(
  UUID = "string",
  FunctionName = "string",
  Enabled = TRUE|FALSE,
  BatchSize = 123
)
}
}

\examples{
# This operation updates a Lambda function event source mapping
\donttest{update_event_source_mapping(
  BatchSize = 123L,
  Enabled = TRUE,
  FunctionName = "myFunction",
  UUID = "1234xCy789012"
)}

}
