% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{update_function_configuration}
\alias{update_function_configuration}
\title{Updates the configuration parameters for the specified Lambda function by using the values provided in the request}
\usage{
update_function_configuration(FunctionName, Role = NULL, Handler = NULL,
  Description = NULL, Timeout = NULL, MemorySize = NULL,
  VpcConfig = NULL, Environment = NULL, Runtime = NULL,
  DeadLetterConfig = NULL, KMSKeyArn = NULL, TracingConfig = NULL,
  RevisionId = NULL)
}
\arguments{
\item{FunctionName}{The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Role}{The Amazon Resource Name (ARN) of the IAM role that Lambda will assume when it executes your function.}

\item{Handler}{The function that Lambda calls to begin executing your function. For Node.js, it is the \code{module-name.export} value in your function.}

\item{Description}{A short user-defined function description. AWS Lambda does not use this value. Assign a meaningful description as you see fit.}

\item{Timeout}{The amount of time that Lambda allows a function to run before terminating it. The default is 3 seconds. The maximum allowed value is 900 seconds.}

\item{MemorySize}{The amount of memory, in MB, your Lambda function is given. AWS Lambda uses this memory size to infer the amount of CPU allocated to your function. Your function use-case determines your CPU and memory requirements. For example, a database operation might need less memory compared to an image processing function. The default value is 128 MB. The value must be a multiple of 64 MB.}

\item{VpcConfig}{Specify security groups and subnets in a VPC to which your Lambda function needs access.}

\item{Environment}{The parent object that contains your environment's configuration settings.}

\item{Runtime}{The runtime version for the function.}

\item{DeadLetterConfig}{A dead letter queue configuration that specifies the queue or topic where Lambda sends asynchronous events when they fail processing. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/dlq.html}{Dead Letter Queues}.}

\item{KMSKeyArn}{The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If you elect to use the AWS Lambda default service key, pass in an empty string ("") for this parameter.}

\item{TracingConfig}{Set \code{Mode} to \code{Active} to sample and trace a subset of incoming requests with AWS X-Ray.}

\item{RevisionId}{An optional value you can use to ensure you are updating the latest update of the function version or alias. If the \code{RevisionID} you pass doesn't match the latest \code{RevisionId} of the function or alias, it will fail with an error message, advising you to retrieve the latest function version or alias \code{RevisionID} using either GetFunction or GetAlias.}
}
\description{
Updates the configuration parameters for the specified Lambda function by using the values provided in the request. You provide only the parameters you want to change. This operation must only be used on an existing Lambda function and cannot be used to update the function's code.
}
\details{
If you are using the versioning feature, note this API will always update the \$LATEST version of your Lambda function. For information about the versioning feature, see \href{http://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html}{AWS Lambda Function Versioning and Aliases}.

This operation requires permission for the \code{lambda:UpdateFunctionConfiguration} action.
}
\examples{
# This operation updates a Lambda function's configuration
update_function_configuration(
  Description = "",
  FunctionName = "myFunction",
  Handler = "index.handler",
  MemorySize = 128L,
  Role = "arn:aws:iam::123456789012:role/lambda_basic_execution",
  Runtime = "python2.7",
  Timeout = 123L,
  VpcConfig = structure(
    list(),
    .Names = character(
      0
    )
  )
)

}
