% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{create_alias}
\alias{create_alias}
\title{Creates an alias that points to the specified Lambda function version}
\usage{
create_alias(FunctionName, Name, FunctionVersion, Description = NULL,
  RoutingConfig = NULL)
}
\arguments{
\item{FunctionName}{[required] The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Name}{[required] Name for the alias you are creating.}

\item{FunctionVersion}{[required] Lambda function version for which you are creating the alias.}

\item{Description}{Description of the alias.}

\item{RoutingConfig}{Specifies an additional version your alias can point to, allowing you to dictate what percentage of traffic will invoke each version. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/lambda-traffic-shifting-using-aliases.html}{Traffic Shifting Using Aliases}.}
}
\description{
Creates an alias that points to the specified Lambda function version. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/aliases-intro.html}{Introduction to AWS Lambda Aliases}.
}
\details{
Alias names are unique for a given function. This requires permission for the lambda:CreateAlias action.
}
\section{Accepted Parameters}{
\preformatted{create_alias(
  FunctionName = "string",
  Name = "string",
  FunctionVersion = "string",
  Description = "string",
  RoutingConfig = list(
    AdditionalVersionWeights = list(
      123.0
    )
  )
)
}
}

