% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{create_function}
\alias{create_function}
\title{Creates a new Lambda function}
\usage{
create_function(FunctionName, Runtime, Role, Handler, Code,
  Description = NULL, Timeout = NULL, MemorySize = NULL,
  Publish = NULL, VpcConfig = NULL, DeadLetterConfig = NULL,
  Environment = NULL, KMSKeyArn = NULL, TracingConfig = NULL,
  Tags = NULL)
}
\arguments{
\item{FunctionName}{[required] The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Runtime}{[required] The runtime version for the function.}

\item{Role}{[required] The Amazon Resource Name (ARN) of the function's \href{http://docs.aws.amazon.com/lambda/latest/dg/intro-permission-model.html#lambda-intro-execution-role}{execution role}.}

\item{Handler}{[required] The name of the method within your code that Lambda calls to execute your function. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/programming-model-v2.html}{Programming Model}.}

\item{Code}{[required] The code for the function.}

\item{Description}{A description of the function.}

\item{Timeout}{The amount of time that Lambda allows a function to run before terminating it. The default is 3 seconds. The maximum allowed value is 900 seconds.}

\item{MemorySize}{The amount of memory that your function has access to. Increasing the function's memory also increases it's CPU allocation. The default value is 128 MB. The value must be a multiple of 64 MB.}

\item{Publish}{Set to true to publish the first version of the function during creation.}

\item{VpcConfig}{If your Lambda function accesses resources in a VPC, you provide this parameter identifying the list of security group IDs and subnet IDs. These must belong to the same VPC. You must provide at least one security group and one subnet ID.}

\item{DeadLetterConfig}{A dead letter queue configuration that specifies the queue or topic where Lambda sends asynchronous events when they fail processing. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/dlq.html}{Dead Letter Queues}.}

\item{Environment}{Environment variables that are accessible from function code during execution.}

\item{KMSKeyArn}{The ARN of the KMS key used to encrypt your function's environment variables. If not provided, AWS Lambda will use a default service key.}

\item{TracingConfig}{Set \code{Mode} to \code{Active} to sample and trace a subset of incoming requests with AWS X-Ray.}

\item{Tags}{The list of tags (key-value pairs) assigned to the new function. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/tagging.html}{Tagging Lambda Functions} in the \strong{AWS Lambda Developer Guide}.}
}
\description{
Creates a new Lambda function. The function configuration is created from the request parameters, and the code for the function is provided by a .zip file. The function name is case-sensitive.
}
\details{
This operation requires permission for the \code{lambda:CreateFunction} action.
}
\section{Accepted Parameters}{
\preformatted{create_function(
  FunctionName = "string",
  Runtime = "nodejs"|"nodejs4.3"|"nodejs6.10"|"nodejs8.10"|"java8"|"python2.7"|"python3.6"|"dotnetcore1.0"|"dotnetcore2.0"|"dotnetcore2.1"|"nodejs4.3-edge"|"go1.x",
  Role = "string",
  Handler = "string",
  Code = list(
    ZipFile = raw,
    S3Bucket = "string",
    S3Key = "string",
    S3ObjectVersion = "string"
  ),
  Description = "string",
  Timeout = 123,
  MemorySize = 123,
  Publish = TRUE|FALSE,
  VpcConfig = list(
    SubnetIds = list(
      "string"
    ),
    SecurityGroupIds = list(
      "string"
    )
  ),
  DeadLetterConfig = list(
    TargetArn = "string"
  ),
  Environment = list(
    Variables = list(
      "string"
    )
  ),
  KMSKeyArn = "string",
  TracingConfig = list(
    Mode = "Active"|"PassThrough"
  ),
  Tags = list(
    "string"
  )
)
}
}

\examples{
# This example creates a Lambda function.
\donttest{create_function(
  Code = structure(
    list(),
    .Names = character(
      0
    )
  ),
  Description = "",
  FunctionName = "MyFunction",
  Handler = "souce_file.handler_name",
  MemorySize = 128L,
  Publish = TRUE,
  Role = "arn:aws:iam::123456789012:role/service-role/role-name",
  Runtime = "nodejs4.3",
  Timeout = 15L,
  VpcConfig = structure(
    list(),
    .Names = character(
      0
    )
  )
)}

}
