% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{delete_alias}
\alias{delete_alias}
\title{Deletes the specified Lambda function alias}
\usage{
delete_alias(FunctionName, Name)
}
\arguments{
\item{FunctionName}{The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Name}{Name of the alias to delete.}
}
\description{
Deletes the specified Lambda function alias. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/aliases-intro.html}{Introduction to AWS Lambda Aliases}.
}
\details{
This requires permission for the lambda:DeleteAlias action.
}
\examples{
# This operation deletes a Lambda function alias
delete_alias(
  FunctionName = "myFunction",
  Name = "alias"
)

}
