% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{delete_function}
\alias{delete_function}
\title{Deletes a Lambda function}
\usage{
delete_function(
  FunctionName = "string",
  Qualifier = "string"
)
}
\arguments{
\item{FunctionName}{[required] The name of the Lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{Qualifier}{Specify a version to delete. You cannot delete a version that is referenced by an alias.}
}
\description{
Deletes a Lambda function. To delete a specific function version, use the \code{Qualifier} parameter. Otherwise, all versions and aliases are deleted. Event source mappings are not deleted.
}
\details{
This operation requires permission for the \code{lambda:DeleteFunction} action.
}
\examples{
# This operation deletes a Lambda function
delete_function(
  FunctionName = "myFunction",
  Qualifier = "1"
)

}
