% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{get_event_source_mapping}
\alias{get_event_source_mapping}
\title{Returns configuration information for the specified event source mapping (see CreateEventSourceMapping)}
\usage{
get_event_source_mapping(UUID)
}
\arguments{
\item{UUID}{The AWS Lambda assigned ID of the event source mapping.}
}
\description{
Returns configuration information for the specified event source mapping (see CreateEventSourceMapping).
}
\details{
This operation requires permission for the \code{lambda:GetEventSourceMapping} action.
}
\examples{
# This operation retrieves a Lambda function's event source mapping
get_event_source_mapping(
  UUID = "123489-xxxxx-kdla8d89d7"
)

}
