% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{get_event_source_mapping}
\alias{get_event_source_mapping}
\title{Returns details about an event source mapping}
\usage{
get_event_source_mapping(
  UUID = "string"
)
}
\arguments{
\item{UUID}{[required] The identifier of the event source mapping.}
}
\description{
Returns details about an event source mapping.
}
\examples{
# This operation retrieves a Lambda function's event source mapping
get_event_source_mapping(
  UUID = "123489-xxxxx-kdla8d89d7"
)

}
