% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{list_aliases}
\alias{list_aliases}
\title{Returns a list of aliases for a Lambda function}
\usage{
list_aliases(FunctionName, FunctionVersion = NULL, Marker = NULL,
  MaxItems = NULL)
}
\arguments{
\item{FunctionName}{[required] The name of the lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} - \code{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \code{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.}

\item{FunctionVersion}{Specify a function version to only list aliases that invoke that version.}

\item{Marker}{Specify the pagination token returned by a previous request to retrieve the next page of results.}

\item{MaxItems}{Limit the number of aliases returned.}
}
\description{
Returns a list of \href{http://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html}{aliases} for a Lambda function.
}
\section{Accepted Parameters}{
\preformatted{list_aliases(
  FunctionName = "string",
  FunctionVersion = "string",
  Marker = "string",
  MaxItems = 123
)
}
}

\examples{
# This operation retrieves a Lambda function's aliases
\donttest{list_aliases(
  FunctionName = "myFunction",
  FunctionVersion = "1",
  Marker = "",
  MaxItems = 123L
)}

}
