% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{remove_layer_version_permission}
\alias{remove_layer_version_permission}
\title{Removes a statement from the permissions policy for a layer version}
\usage{
remove_layer_version_permission(
  LayerName = "string",
  VersionNumber = 123,
  StatementId = "string",
  RevisionId = "string"
)
}
\arguments{
\item{LayerName}{[required] The name of the layer.}

\item{VersionNumber}{[required] The version number.}

\item{StatementId}{[required] The identifier that was specified when the statement was added.}

\item{RevisionId}{Only update the policy if the revision ID matches the ID specified. Use this option to avoid modifying a policy that has changed since you last read it.}
}
\description{
Removes a statement from the permissions policy for a layer version. For more information, see AddLayerVersionPermission.
}
\examples{

}
