% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lambda_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Creates a list of tags (key-value pairs) on the Lambda function}
\usage{
tag_resource(
  Resource = "string",
  Tags = list(
    "string"
  )
)
}
\arguments{
\item{Resource}{[required] The ARN (Amazon Resource Name) of the Lambda function. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/tagging.html}{Tagging Lambda Functions} in the \strong{AWS Lambda Developer Guide}.}

\item{Tags}{[required] The list of tags (key-value pairs) you are assigning to the Lambda function. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/tagging.html}{Tagging Lambda Functions} in the \strong{AWS Lambda Developer Guide}.}
}
\description{
Creates a list of tags (key-value pairs) on the Lambda function. Requires the Lambda function ARN (Amazon Resource Name). If a key is specified without a value, Lambda creates a tag with the specified key and a value of null. For more information, see \href{http://docs.aws.amazon.com/lambda/latest/dg/tagging.html}{Tagging Lambda Functions} in the \strong{AWS Lambda Developer Guide}.
}
\examples{

}
